extract C Bindings
