+++ skia-bindings/build_support.rs
@@ -6,6 +6,7 @@ pub mod clang;
 pub mod ios;
 pub mod llvm;
 pub mod skia;
+pub mod skia_c_bindings;
 pub mod vs;
 pub mod xcode;
